+++ back-end/api-service/routes/status.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/status', (req, res) => {
+    res.status(200).json({ status: 'API is running!' });
+});
+
+module.exports = router;
