+++ back-end/api-service/routes/users.js
@@ -1,0 +1,9 @@
+// users.js
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res, next) => {
+    res.json({ message: 'Users route' });
+});
+
+module.exports = router;
