+++ back-end/api-service/views/index.pug
@@ -1,0 +1,6 @@
+doctype html
+html
+    head
+        title Video Transcoder
+    body
+        h1 Welcome to Video Transcoder
