+++ back-end/app.js
@@ -2,75 +2,69 @@ require('dotenv').config();
 const express = require('express');
 const bodyParser = require('body-parser');
 const path = require('path');
-// routes
-const indexRouter = require('./routes/index');
-const authRouter = require('./routes/auth');
-const videoRouter = require('./routes/video');
-const pexelsRouter = require('./routes/pexels');
-
-const { setupDynamoDBTable } = require('./models/dynamoDBClient'); // Import table setup
-
-// cors (for inner network/ different ports)
 const cors = require('cors');
 
-// others
+// Import routes
+const indexRouter = require('./api-service/routes/index');
+const authRouter = require('./api-service/routes/auth');
+const videoRouter = require('./api-service/routes/video');
+const pexelsRouter = require('./api-service/routes/pexels');
 
-const { verifyToken } = require('./controllers/authController');
+// Import setup and middleware
+const { setupDynamoDBTable } = require('./api-service/models/dynamoDBClient');
+const { verifyToken } = require('./api-service/controllers/authController');
 
+// Initialize Express app
 const app = express();
+
+// Configure CORS
 const corsOptions = {
     origin: process.env.ALLOWED_ORIGIN || 'http://n10937668.cab432.com:3000',
     methods: ['GET', 'POST', 'PUT', 'DELETE'],
     allowedHeaders: ['Content-Type', 'Authorization', 'x-access-token'],
     credentials: true,
 };
-
 app.use(cors(corsOptions));
 
+// Setup public directory for static files
 app.use('/uploads', express.static(path.join(__dirname, '../uploads')));
+app.use(express.static(path.join(__dirname, 'public')));
 
-// mongoose.connect(process.env.MONGODB_URI, {
-// useNewUrlParser: true,
-//     useUnifiedTopology: true,
-//     ssl: false,
-//     tls: false// Disable SSL
-// })
-// .then(() => console.log('MongoDB Connected'))
-//     .catch(err => console.log(err));
+// Initialize DynamoDB table setup
+setupDynamoDBTable()
+    .then(() => console.log("DynamoDB table is ready."))
+    .catch(err => console.error("Error setting up DynamoDB table:", err));
 
-setupDynamoDBTable().then(() => {
-    console.log("DynamoDB table is ready.");
-}).catch(err => {
-    console.error("Error setting up DynamoDB table:", err);
-});
-
-
-app.set('views', path.join(__dirname, 'views'));
+// View engine setup
+app.set('views', path.join(__dirname, 'api-service','views'));
 app.set('view engine', 'pug');
 
+// Middleware for parsing requests
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
-app.use(express.static(path.join(__dirname, 'public')));
+
+// Route registration
 app.use('/', indexRouter);
 app.use('/auth', authRouter);
-// app.use('/video', verifyToken, videoRouter);
-app.use('/video', videoRouter);
+app.use('/video', videoRouter); // Optional token verification disabled for now
 app.use('/pexels', verifyToken, pexelsRouter);
 
-
-
+// Handle 404 errors (unknown routes)
 app.use((req, res, next) => {
-    const err = new Error('Not Found');
-    err.status = 404;
-    next(err);
+    const error = new Error('Not Found');
+    error.status = 404;
+    next(error);
 });
 
-app.use((err, req, res, next) => {
-    res.status(err.status || 500);
-    res.json({ error: err.message });
+// Error handler middleware
+app.use((error, req, res, next) => {
+    res.status(error.status || 500).json({ error: error.message });
 });
+
+// Serve index.html for all unmatched routes (Single Page Application behavior)
 app.get('*', (req, res) => {
     res.sendFile(path.join(__dirname, 'public', 'index.html'));
 });
 
+// Export the app for use in bin/www.js
 module.exports = app;
