+++ back-end/package.json
@@ -33,6 +33,7 @@
     "ffprobe-static": "^3.1.0",
     "fluent-ffmpeg": "^2.1.3",
     "multer": "^1.4.5-lts.1",
+    "pug": "^3.0.3",
     "qrcode": "^1.5.4",
     "stream": "^0.0.3"
   }
