--- back-end/routes/index.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('/', (req, res) => {
-    res.render('index', { title: 'Video Transcoder' });
-});
-
-module.exports = router;
