--- back-end/routes/users.js
@@ -1,9 +1,0 @@
-// users.js
-const express = require('express');
-const router = express.Router();
-
-router.get('/', (req, res, next) => {
-    res.json({ message: 'Users route' });
-});
-
-module.exports = router;
