+++ back-end/video-service/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "video-service",
+    "version": "1.0.0",
+    "description": "Microservice for video transcoding",
+    "main": "videoProcessor.js",
+    "scripts": {
+        "start": "node videoProcessor.js"
+    },
+    "dependencies": {
+        "@aws-sdk/client-s3": "^3.679.0",
+        "@aws-sdk/s3-request-presigner": "^3.679.0",
+        "aws-sdk": "^2.1190.0",
+        "dotenv": "^16.4.5",
+        "express": "^4.19.2",
+        "ffmpeg-static": "^5.2.0",
+        "ffprobe-static": "^3.1.0",
+        "multer": "^1.4.5-lts.1"
+    }
+}
