--- back-end/views/index.pug
@@ -1,6 +1,0 @@
-doctype html
-html
-    head
-        title Video Transcoder
-    body
-        h1 Welcome to Video Transcoder
